fix: add max retry handling
