+++ src/ETransferServer.Grains/State/Order/WithdrawOrderCallState.cs
@@ -7,4 +7,5 @@ public class WithdrawOrderCallState
     [Id(1)] public int Status { get; set; }
     [Id(2)] public int CallRetry { get; set; }
     [Id(3)] public int CallbackRetry { get; set; }
+    [Id(4)] public int CallQueryRetry { get; set; }
 }
